\name{dataTOY}
\alias{dataTOY}
\docType{data}
\title{
Artificial data set
}
\description{
Artificial multilevel data set.

}
\usage{data("dataTOY")}
\format{
  A data frame with 3000 observations on the following 13 variables.
  \describe{
    \item{\code{id_high}}{High-level id}
    \item{\code{Y_1}}{Indicator n.1}
    \item{\code{Y_2}}{Indicator n.2}
    \item{\code{Y_3}}{Indicator n.3}
    \item{\code{Y_4}}{Indicator n.4}
    \item{\code{Y_5}}{Indicator n.5}
    \item{\code{Y_6}}{Indicator n.6}
    \item{\code{Y_7}}{Indicator n.7}
    \item{\code{Y_8}}{Indicator n.8}
    \item{\code{Y_9}}{Indicator n.9}
    \item{\code{Y_10}}{Indicator n.10}
    \item{\code{Z_low}}{Continuous low-level covariate}
    \item{\code{Z_high}}{Continuous high-level covariate}
  }
}
\references{
Di Mari, Bakk, Z., R., Oser, J., & Kuha, J. (2023). A two-step estimator for multilevel latent class analysis with covariates. Under review. Available from \url{https://arxiv.org/abs/2303.06091}.
}
\keyword{datasets}